backend/requirements.txt
fastapi
uvicorn[standard]
motor

pymongo
passlib[bcrypt]==1.7.4
python-jose[cryptography]
email-validator
bcrypt==4.0.1
